Add files via upload
This code translates some of the SquareMatrix interface shown in the lectures from C# into Python.  It is not complete but runs. Essentially, it specifies the attributes and methods of any objects (e.g. Gates) which inherit from it.

Instead of a the MatrixElement struct this code uses a class of the same name. There is scope for importing the struct module however I do not believe this is necessary in our case but may be worth investigating.

See the comments for further information.